+++ lib/src/graph/strategy/default.dart
@@ -1,0 +1,9 @@
+import '../strategy.dart';
+
+class DefaultGraphStrategy<T> implements GraphStrategy<T> {
+  @override
+  Set<T> createSet() => <T>{};
+
+  @override
+  Map<T, R> createMap<R>() => <T, R>{};
+}